+++ src/CucumberJson.ts
@@ -1,81 +1,87 @@
 export type CucumberJson = {
   implementation: string
-  features: readonly Feature[]
+  features: readonly CucumberJsonFeature[]
 }
 
-export type Feature = Readonly<{
+export type CucumberJsonFeature = Readonly<{
   uri: string
   id?: string
   line?: number
   keyword: string
   name: string
   description?: string
-  elements: readonly Element[]
-  tags?: readonly Tag[]
+  elements: readonly CucumberJsonElement[]
+  tags?: readonly CucumberJsonTag[]
 }>
 
-export type Element = Readonly<{
+export type CucumberJsonElement = Readonly<{
   start_timestamp?: string
   line: number
   id?: string
-  type: ElementType
+  type: CucumberJsonElementType
   keyword: string
   name: string
   description: string
-  before?: readonly Hook[]
-  steps: readonly Step[]
-  after?: readonly Hook[]
-  tags?: readonly Tag[]
+  before?: readonly CucumberJsonHook[]
+  steps: readonly CucumberJsonStep[]
+  after?: readonly CucumberJsonHook[]
+  tags?: readonly CucumberJsonTag[]
 }>
 
-export type ElementType = 'background' | 'scenario'
+export type CucumberJsonElementType = 'background' | 'scenario'
 
-export type Hook = Readonly<{
-  match?: Match
-  result: Result
+export type CucumberJsonHook = Readonly<{
+  match?: CucumberJsonMatch
+  result: CucumberJsonResult
 }>
 
-export type Step = Readonly<{
+export type CucumberJsonStep = Readonly<{
   keyword: string
   line: number
-  match?: Match
+  match?: CucumberJsonMatch
   name: string
-  result: Result
-  doc_string?: DocString
-  rows?: readonly DataTableRow[]
+  result: CucumberJsonResult
+  doc_string?: CucumberJsonDocString
+  rows?: readonly CucumberJsonDataTableRow[]
 }>
 
-export type DataTableRow = Readonly<{
+export type CucumberJsonDataTableRow = Readonly<{
   cells: readonly string[]
 }>
 
-export type DocString = Readonly<{
+export type CucumberJsonDocString = Readonly<{
   line: number
   value: string
   content_type?: string
 }>
 
-export type Match = Readonly<{
+export type CucumberJsonMatch = Readonly<{
   location?: string
-  arguments?: readonly Argument[]
+  arguments?: readonly CucumberJsonArgument[]
 }>
 
-export type Argument = Readonly<{
+export type CucumberJsonArgument = Readonly<{
   value: string
   offset: number
 }>
 
-export type Result = Readonly<{
+export type CucumberJsonResult = Readonly<{
   duration?: number
-  status: Status
+  status: CucumberJsonStatus
   error_message?: string
 }>
 
 // The unknown value is used if the original result was missing. This is the case for e.g.
 // Scenario Outline steps on Cucumber-Ruby 2.0.0
-export type Status = 'passed' | 'failed' | 'skipped' | 'undefined' | 'pending' | 'unknown'
+export type CucumberJsonStatus =
+  | 'passed'
+  | 'failed'
+  | 'skipped'
+  | 'undefined'
+  | 'pending'
+  | 'unknown'
 
-export type Tag = Readonly<{
+export type CucumberJsonTag = Readonly<{
   name: string
   line?: number
 }>